Split oversized settings regression spec
Move analysis and language browser coverage into focused settings regression specs and remove the settings oversized allowlist entry.
